requirements.txt
# Core data science libraries (updated versions for Python 3.12)
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
xgboost>=1.7.0

# API and deployment
flask>=2.3.0
gunicorn>=20.1.0
docker>=6.1.0

# Data collection
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
jupyter>=1.0.0
notebook>=6.5.0
ipython>=8.14.0
pyyaml>=6.0

# Build dependencies (required for some packages)
setuptools>=65.0.0
wheel>=0.40.0